try to keep last zoom level; adjust zoom level based on location searched ie street, city, state, country

+++ dotnet/PickupGames/PickupGames/Repositories/GoogleGeographyRepository.cs
@@ -43,7 +43,7 @@ namespace PickupGames.Repositories
 
             return new Distance
             {
-                Value = decimal.Parse(array[0]),
+                Value = double.Parse(array[0]),
                 Unit = array[1]
             };
         }
